+++ src/Response/FileResponse.ts
@@ -1,8 +1,8 @@
 import {Response} from './Response';
 
 export class FileResponse extends Response {
-  constructor(protected filename: string, statusCode: number = 200, headers = []) {
-    super(filename, statusCode, headers);
+  constructor(protected filename: string, statusCode: number = 200, headers: Record<string, string | string[]> = {}) {
+    super(null, statusCode, headers);
   }
 
   public getFilename() {