+++ src/Response/XmlFromJsonResponse.ts
@@ -2,10 +2,8 @@ import {Response} from './Response';
 import * as jsonxml from 'jsontoxml';
 
 export class XmlFromJsonResponse extends Response {
-  constructor(protected content = {}, protected statusCode: number = 200, protected headers = []) {
-    super(content, statusCode, headers);
-
-    this.headers.push('Content-Type: text/xml');
+  constructor(content = {}, statusCode: number = 200, headers: Record<string, string | string[]> = {}) {
+    super(content, statusCode, {'Content-Type': 'text/xml', ...headers});
   }
 
   public getContent() {