+++ test/Response/JsonRespose.test.ts
@@ -2,11 +2,22 @@ import {JsonResponse} from '../../src/Response';
 
 describe("JsonResponse", () => {
   test("getters", () => {
-    const response = new JsonResponse({foo: 'bar'}, 400);
+    const response = new JsonResponse({foo: 'bar'}, 400, {'X-Test': 'foobar'});
 
     expect(response.getContent()).toEqual({foo: 'bar'});
     expect(response.getStatusCode()).toEqual(400);
-    expect(response.getHeaders()).toEqual(['Content-Type: application/json']);
+    expect(response.getHeaders()).toEqual({
+      'X-Test': 'foobar',
+      'Content-Type': 'application/json'
+    });
+  });
+
+  test("getters with custom header Content-Type", () => {
+    const response = new JsonResponse("test", 200, {'Content-Type': 'application/ld+json'});
+
+    expect(response.getContent()).toEqual("test");
+    expect(response.getStatusCode()).toEqual(200);
+    expect(response.getHeaders()).toEqual({'Content-Type': 'application/ld+json'});
   });
 
   test("getters with default status", () => {
@@ -14,7 +25,7 @@ describe("JsonResponse", () => {
 
     expect(response.getContent()).toEqual("test");
     expect(response.getStatusCode()).toEqual(200);
-    expect(response.getHeaders()).toEqual(['Content-Type: application/json']);
+    expect(response.getHeaders()).toEqual({'Content-Type': 'application/json'});
   });
 
   test("getters with default content", () => {
@@ -22,6 +33,6 @@ describe("JsonResponse", () => {
 
     expect(response.getContent()).toEqual({});
     expect(response.getStatusCode()).toEqual(200);
-    expect(response.getHeaders()).toEqual(['Content-Type: application/json']);
+    expect(response.getHeaders()).toEqual({'Content-Type': 'application/json'});
   });
 });